requirements.txt
pymongo[snappy,zstd,encryption]
redis-om
redis[asyncio]
aiohttp[speedups]
tortoise-orm[asyncpg]

python-telegram-bot
python-dotenv
pydantic
pydantic-settings
pytz